# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/products":{"post":{"operationId":"ProductsController_createProduct","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}},"responses":{"201":{"description":""}},"tags":["Products"],"security":[{"JWT-auth":[]}]},"get":{"operationId":"ProductsController_getAllProducts","parameters":[],"responses":{"200":{"description":""}},"tags":["Products"],"security":[{"JWT-auth":[]}]}},"/products/{id}":{"get":{"operationId":"ProductsController_getProductById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Products"],"security":[{"JWT-auth":[]}]},"put":{"operationId":"ProductsController_updateProduct","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}},"responses":{"200":{"description":""}},"tags":["Products"],"security":[{"JWT-auth":[]}]},"delete":{"operationId":"ProductsController_deleteProduct","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Products"],"security":[{"JWT-auth":[]}]}},"/autho/signup":{"post":{"operationId":"AuthController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"201":{"description":""}},"tags":["authentication "],"security":[{"JWT-auth":[]}]}},"/autho/signin":{"post":{"operationId":"AuthController_signip","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninDto"}}}},"responses":{"201":{"description":""}},"tags":["authentication "],"security":[{"JWT-auth":[]}]}},"/users/add":{"post":{"operationId":"UsersController_createNewUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"],"security":[{"JWT-auth":[]}]}},"/users/{id}":{"get":{"operationId":"UsersController_findUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"JWT-auth":[]}]},"put":{"operationId":"UsersController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/updateuserDTO"}}}},"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"JWT-auth":[]}]},"delete":{"operationId":"UsersController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"JWT-auth":[]}]}},"/users":{"get":{"operationId":"UsersController_findAllUsers","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"JWT-auth":[]}]}},"/orders":{"post":{"operationId":"OrdersController_createProduct","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Orderdto"}}}},"responses":{"201":{"description":""}},"tags":["Orders"],"security":[{"JWT-auth":[]}]},"get":{"operationId":"OrdersController_getAllProducts","parameters":[],"responses":{"200":{"description":""}},"tags":["Orders"],"security":[{"JWT-auth":[]}]}},"/orders/{id}":{"get":{"operationId":"OrdersController_getProductById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Orders"],"security":[{"JWT-auth":[]}]},"put":{"operationId":"OrdersController_updateProduct","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Orderdto"}}}},"responses":{"200":{"description":""}},"tags":["Orders"],"security":[{"JWT-auth":[]}]},"delete":{"operationId":"OrdersController_deleteProduct","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Orders"],"security":[{"JWT-auth":[]}]}}},"info":{"title":"nest js project","description":"API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"JWT-auth":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ProductDto":{"type":"object","properties":{"namepro":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"orders":{"type":"array","items":{"type":"string"}}},"required":["namepro","description","price","orders"]},"UserDto":{"type":"object","properties":{"name":{"type":"string"},"age":{"type":"number"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"}},"required":["name","age","email","password","role"]},"SigninDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"}},"required":["email","password","role"]},"updateuserDTO":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"}},"required":["name","email","role"]},"Orderdto":{"type":"object","properties":{"customerName":{"type":"string"},"orderDate":{"format":"date-time","type":"string"},"products":{"type":"object"}},"required":["customerName","orderDate","products"]}}}}